Honor bounded FRAM example transport and harden HIL evidence
Use bounded open-drain bus recovery, honor and restore callback timeouts, and avoid diagnostic arithmetic overflow and destructive console drains. Require complete framed payloads and full health evidence in the maintained host runner.

Validation: working native 186/186; isolated staged-only native 183/183, host 18/18 and parser checks; S2/S3 builds. Three MB85RC64TA boards each passed 34 functional checks and a 40-minute soak (6033/6055/6075 commands), full-capacity four-pattern verification, combined verified-write ambiguity/reconciliation, controlled absent-address recovery, and exact original full-memory restoration.

Pre-existing unrelated local changes remain outside this commit.

+++ examples/01_basic_bringup_cli/main.cpp
@@ -173,7 +173,7 @@ void printDriverHealth() {
   const uint32_t now = millis();
   const uint32_t totalOk = device.totalSuccess();
   const uint32_t totalFail = device.totalFailures();
-  const uint32_t total = totalOk + totalFail;
+  const uint64_t total = static_cast<uint64_t>(totalOk) + totalFail;
   const float successRate = (total > 0U)
                                 ? (100.0f * static_cast<float>(totalOk) / static_cast<float>(total))
                                 : 0.0f;
